# src/Api/bin/Debug/net6.0/Swagger/OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pegi",
    "description": "Es una api para la gestion de proyectos de investigación",
    "contact": {
      "name": "Ana Patiño",
      "email": "sofipatino614@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/Professor": {
      "post": {
        "tags": [
          "Professor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Professor"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Professor/{document}": {
      "get": {
        "tags": [
          "Professor"
        ],
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Professor/get-professor-position/{position}": {
      "get": {
        "tags": [
          "Professor"
        ],
        "parameters": [
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProfessorRequest": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}